test(quickadd): move the panel-refusal cases to the real-panel target
Cloud review, adopted. `present` builds its panel through `ensurePanel()` BEFORE it
measures anything, so any test reaching the refusal puts a window-server-backed
`NSPanel` into the process. The injected presenter replaces the KEYING and not the
CONSTRUCTION, so "no window is ordered on screen" was true and "no real panel is
created" was not.

`EnviousWisprDesktopEffectsTests` is the target for exactly this — it already holds
`OverlayRetainedWindowRealPanelTests` and `RecordingPollingRealPanelTests`, carries
`@testable import EnviousWisprAppKit`, and runs in the same Debug scheme. Checked
rather than taken on report: `Project.swift:621` declares it, and the two files above
are the precedent.

So the cases move rather than being weakened, and the unit lane stays free of live
desktop effects.

A local `RecordingPresenter` because the unit target's recorder is not visible from
here. It answers the single member `present` reaches.

Re-proven WHERE IT LANDED, since a test that moved has to be re-run in its new home:

  baseline before  2 tests green
  [ EXP ] row 1    exact must_fire set matched (anUnmeasurablePanelIsRefused)
                   1 named silent control unchanged (aMeasurablePanelStillOpens)
  baseline after   2 tests green, tree restored

Filter is now `EnviousWisprDesktopEffectsTests/QuickAddPanelHostRealPanelTests`.
#2542 left the suite filter open ("or wherever the new case lands"), so the frozen
row is honoured rather than edited.

Part of #2542.

Self-audit-grep: /usr/bin/grep -rn "QuickAddPanelHost" Tests/ Sources/
Self-audit-owner: Tests/EnviousWisprDesktopEffectsTests/QuickAddPanelHostRealPanelTests.swift
